+++ dev/Unibook/UsersDal/UserDal.cs
@@ -78,7 +78,7 @@ namespace Univalle.Fie.Sistemas.UniBook.UsersDal
         /// </summary>
         /// <param name="id"></param>
         /// <param name="objContex"></param>
-        public static void Delete(int id, ModelUnibookContainer objContex)
+        public static void Delete(long id, ModelUnibookContainer objContex)
         {
             try
             {
